requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.22
alembic==1.11.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.29.0
tweepy==4.16.0
google-api-python-client==2.99.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.1
python-dotenv==1.0.1
requests==2.31.0
scikit-learn==1.3.1
prophet==1.1.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytest==7.4.2
